requirements.txt
# Core
pytz==2022.1
django==3.2.13
django-extensions==3.1.5
django-configurations==2.4.1
gunicorn==20.1.0

# For the persistence stores
dj-database-url==2.0.0
psycopg2-binary==2.9.3
# psycopg2==2.9.3

# Model Tools
django-model-utils==4.1.1
django-cors-headers==4.3.1

# Rest apis
djangorestframework==3.12.4

# Developer Tools
ipdb==0.13.9
mkdocs==1.4.3
flake8==4.0.1

# Testing
pytest-mock==3.6.1
factory-boy==3.2.0
coverage==6.1.2

# Static and Media Storage
boto3==1.17.54